write assertion test for pattern library and util library
write assertion test for pattern_library.js and pattern_util.js

modify the run_all_test.sh shell script to run the assertion test of the
two libraries and include a logic in it which catches the failing tests
and gives back a simple report.

next motive is to remove the process.argv logic from the mains and put
them into libraries as well, making the main files totally logic-free.